NMS-14541 Bridge Topology Discovery, Move BridgeTopologyException
Move static create List bridgeMacLinks
ro BridgeTopologyService
