+++ features/enlinkd/service/api/src/main/java/org/opennms/netmgt/enlinkd/service/api/BridgePortWithMacs.java
@@ -53,7 +53,7 @@ public class BridgePortWithMacs implements Topology {
     }
 
     public List<BridgeMacLink> getBridgeMacLinks() {
-        return TopologyService.create(m_port, m_macs, BridgeMacLink.BridgeMacLinkType.BRIDGE_FORWARDER);
+        return BridgeTopologyService.create(m_port, m_macs, BridgeMacLink.BridgeMacLinkType.BRIDGE_FORWARDER);
     }
 
     @Override
