+++ features/enlinkd/service/api/src/main/java/org/opennms/netmgt/enlinkd/service/api/BridgeTopologyService.java
@@ -29,16 +29,35 @@
 package org.opennms.netmgt.enlinkd.service.api;
 
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
 import org.opennms.netmgt.enlinkd.model.BridgeElement;
+import org.opennms.netmgt.enlinkd.model.BridgeMacLink;
 import org.opennms.netmgt.enlinkd.model.BridgeStpLink;
+import org.opennms.netmgt.model.OnmsNode;
 
 public interface BridgeTopologyService extends TopologyService {
 
+    static List<BridgeMacLink> create(BridgePort bp, Set<String> macs, BridgeMacLink.BridgeMacLinkType type) {
+        final List<BridgeMacLink> maclinks = new ArrayList<>();
+        macs.forEach(mac -> {
+            BridgeMacLink maclink = new BridgeMacLink();
+            OnmsNode node = new OnmsNode();
+            node.setId(bp.getNodeId());
+            maclink.setNode(node);
+            maclink.setBridgePort(bp.getBridgePort());
+            maclink.setBridgePortIfIndex(bp.getBridgePortIfIndex());
+            maclink.setMacAddress(mac);
+            maclink.setVlan(bp.getVlan());
+            maclink.setLinkType(type);
+            maclinks.add(maclink);
+        });
+        return maclinks;
+    }
     // this indicates the total size of in memory bft
     boolean collectBft(int nodeid, int maxsize);
